startTournament clean
